+++ extensions/iceworks-style-helper/src/cssClassAutoCompete/index.ts
@@ -1,7 +1,7 @@
 import * as fs from 'fs';
 import * as path from 'path';
 import * as vscode from 'vscode';
-import { recordDAU } from '@iceworks/recorder';
+import { recordCompletionItemProvider } from '@iceworks/recorder';
 import { getFocusCodeInfo } from '../getFocusCodeInfo';
 
 function unique(arr: string[]) {
@@ -27,10 +27,14 @@ function provideCompletionItems(document: vscode.TextDocument, position: vscode.
     }
   });
   if (classNames.length) {
-    recordDAU();
+    recordCompletionItemProvider();
   }
   return unique(classNames).map((className) => {
-    return new vscode.CompletionItem(`.${className}`, vscode.CompletionItemKind.Text);
+    const completionItem = new vscode.CompletionItem(`.${className}`, vscode.CompletionItemKind.Text);
+    completionItem.detail = 'Iceworks';
+    completionItem.insertText = `.${className} {\n  \n}`;
+    completionItem.command = { command: 'iceworksApp.recorder.recordCompletionItemSelect', title: '' };
+    return completionItem;
   });
 }